// AETHERION/frontend/options-analytics-dashboard/utils/types.ts
export interface OptionRow {
  symbol: string;
  datetime: string;
  expiry: string;
  CE: number;
  PE: number;
  spot_close: number;
  ATM: number;
  strike: number;
  oi_CE: number;
  oi_PE: number;
  volume_CE: number;
  volume_PE: number;
}